(bluefox) implement start/stop of ioBroker
