+++ .gitignore
@@ -8,3 +8,7 @@ backitup.sh.lastrun
 backitupl.sh
 package-lock.json
 /package-lock.json
+/lib/startIOB.bat
+/lib/*.bat
+/lib/*.log
+/lib/total.json
