+++ lib/tools.js
@@ -37,7 +37,23 @@ function copyFile(source, target, cb) {
     rd.pipe(wr);
 }
 
+/**
+ * looks for iobroker home folder
+ *
+ * @returns {string}
+ */
+function getIobDir() {
+    const utils = require('./utils');
+    const tools = require(utils.controllerDir + '/lib/tools.js');
+    let backupDir = tools.getConfigFileName().replace(/\\/g, '/');
+    let parts = backupDir.split('/');
+    parts.pop(); // iobroker.json
+    parts.pop(); // iobroker-data.json
+    return parts.join('/');
+}
+
 module.exports = {
     getDate,
-    copyFile
+    copyFile,
+    getIobDir
 };
