+++ package.json
@@ -17,6 +17,10 @@
     {
       "name": "peoples0815",
       "email": "peoples0815github@gmail.com"
+    },
+    {
+      "name": "bluefox",
+      "email": "dogafox@gmail.com"
     }
   ],
   "homepage": "https://github.com/simatec/ioBroker.backitup",
